Move request context overloads to ASP.NET Core
